Update syslog-ng configuration if needed from configure-lightwave-server script. Check for existence of syslog-ng.conf before updating.

+++ lw-server/package/rpm/vmware-lightwave-server.spec
@@ -25,14 +25,17 @@ VMware Infrastructure Controller
     # First argument is 2 => Upgrade
 case "$1" in
     1)
+        # Configure syslog-ng
         LINE='@include "lightwave.conf.d"'
         FILE=/etc/syslog-ng/syslog-ng.conf
-        grep -q "$LINE" "$FILE"
-        if [ "$?" -ne 0 ]; then
-            echo "$LINE" >> "$FILE"
-            pid=$( pidof syslog-ng )
-            if [ -n "$pid" ]; then
-                kill -HUP $pid
+        if [ -f "$FILE" ]; then
+            grep -qs "$LINE" "$FILE"
+            if [ "$?" -ne 0 ]; then
+                echo "$LINE" >> "$FILE"
+                pid=$( pidof syslog-ng )
+                if [ -n "$pid" ]; then
+                    kill -HUP $pid
+                fi
             fi
         fi
         ;;
